[FEATURE] Common utils package
