+++ apps/web/package.json
@@ -17,6 +17,7 @@
   },
   "dependencies": {
     "@backtrade/types": "workspace:*",
+    "@backtrade/utils": "workspace:*",
     "@tanstack/react-query": "^5.90.7",
     "lightweight-charts": "^5.0.9",
     "react": "^19.2.0",
