+++ apps/web/src/features/session/session-running/components/PositionDetailsModal/PositionDetailsModal.tsx
@@ -1,6 +1,7 @@
 import { useEffect } from "react";
 import { Button } from "../../../../../components";
 import type { Position } from "@backtrade/types";
+import { formatDateTime } from "@backtrade/utils";
 import styles from "./PositionDetailsModal.module.css";
 
 interface PositionDetailsModalProps {
@@ -45,18 +46,6 @@ export function PositionDetailsModal({
   }, [isOpen]);
 
   if (!isOpen || !position) return null;
-
-  const formatDateTime = (dateString: string): string => {
-    const date = new Date(dateString);
-    return date.toLocaleString("en-US", {
-      year: "numeric",
-      month: "short",
-      day: "numeric",
-      hour: "2-digit",
-      minute: "2-digit",
-      second: "2-digit",
-    });
-  };
 
   return (
     <div className={styles.backdrop} onClick={onClose}>
