+++ apps/web/src/features/settings/utils/validation.ts
@@ -1,73 +1,15 @@
 /**
- * Email validation result
+ * Re-export shared validation utilities from @backtrade/utils
+ * This file is kept for backward compatibility with existing imports.
+ * New code should import directly from @backtrade/utils.
  */
-export interface EmailValidationResult {
-  isValid: boolean;
-  error?: string;
-}
-
-/**
- * Password validation result
- */
-export interface PasswordValidationResult {
-  isValid: boolean;
-  error?: string;
-}
-
-/**
- * Validate email format
- *
- * @param email - Email address to validate
- * @returns Validation result
- */
-export function validateEmail(email: string): EmailValidationResult {
-  if (!email || email.trim() === "") {
-    return { isValid: false, error: "Email is required" };
-  }
-
-  const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
-  if (!emailRegex.test(email)) {
-    return { isValid: false, error: "Invalid email format" };
-  }
-
-  return { isValid: true };
-}
-
-/**
- * Validate password strength
- *
- * @param password - Password to validate
- * @returns Validation result
- */
-export function validatePassword(password: string): PasswordValidationResult {
-  if (!password || password.length === 0) {
-    return { isValid: false, error: "Password is required" };
-  }
-
-  if (password.length < 8) {
-    return {
-      isValid: false,
-      error: "Password must be at least 8 characters",
-    };
-  }
-
-  return { isValid: true };
-}
-
-/**
- * Validate password confirmation
- *
- * @param password - Original password
- * @param confirmPassword - Confirmation password
- * @returns Validation result
- */
-export function validatePasswordConfirmation(
-  password: string,
-  confirmPassword: string,
-): PasswordValidationResult {
-  if (password !== confirmPassword) {
-    return { isValid: false, error: "Passwords don't match" };
-  }
-
-  return validatePassword(password);
-}
+export type {
+  ValidationResult,
+  EmailValidationResult,
+  PasswordValidationResult,
+} from "@backtrade/utils";
+export {
+  validateEmail,
+  validatePassword,
+  validatePasswordConfirmation,
+} from "@backtrade/utils";
