+++ packages/utils/eslint.config.js
@@ -1,0 +1,13 @@
+import baseConfig from "@backtrade/eslint-config";
+
+export default [
+  ...baseConfig,
+  {
+    languageOptions: {
+      parserOptions: {
+        tsconfigRootDir: import.meta.dirname,
+        project: "./tsconfig.json",
+      },
+    },
+  },
+];
