+++ packages/utils/package.json
@@ -1,0 +1,34 @@
+{
+  "name": "@backtrade/utils",
+  "version": "1.0.0",
+  "private": true,
+  "type": "module",
+  "main": "./src/index.ts",
+  "types": "./src/index.ts",
+  "exports": {
+    ".": {
+      "types": "./src/index.ts",
+      "import": "./src/index.ts",
+      "require": "./src/index.ts"
+    }
+  },
+  "scripts": {
+    "dev": "tsc --watch --project tsconfig.json",
+    "build": "tsc --project tsconfig.json",
+    "lint": "eslint . --ext .ts,.tsx",
+    "lint:fix": "eslint . --ext .ts,.tsx --fix",
+    "typecheck": "tsc --noEmit"
+  },
+  "dependencies": {
+    "@backtrade/types": "workspace:*"
+  },
+  "license": "PROPRIETARY",
+  "author": "REICHHART Damien",
+  "devDependencies": {
+    "@backtrade/eslint-config": "workspace:*",
+    "@backtrade/tsconfig": "workspace:*",
+    "globals": "^16.5.0",
+    "typescript": "^5.9.3",
+    "typescript-eslint": "8.46.4"
+  }
+}
