+++ packages/utils/src/index.ts
@@ -1,0 +1,7 @@
+/**
+ * Shared utilities for BackTrade
+ * This package contains common utilities that can be used in both frontend and backend.
+ */
+
+export * from "./time";
+export * from "./validation";
